% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT_timeshift.R
\name{calcColorSumStats}
\alias{calcColorSumStats}
\title{Function to perfomr summary statistic for each image}
\usage{
calcColorSumStats(img_path, nrow = 9, ncol = 16)
}
\arguments{
\item{img_path}{img_path this the folder directory where the image are located}

\item{nrow}{A grid number of row}

\item{ncol}{A grid number of colum}
}
\value{
return a dataframe with proportion of black and white or color for each imagesy
}
\description{
Function to perfomr summary statistic for each image
}
\keyword{internal}
