% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{CTTimeShift-package}
\alias{CTTimeShift}
\alias{CTTimeShift-package}
\title{CTTimeShift: Analyze Camera Trap Image Metadata and Color Patterns}
\description{
This package provides tools to extract metadata and perform grid-based color analysis by clasifying each images in grayscale or color using Kmeans clustering to identify if an image is a day or nigth
on camera trap images, helping to correct timestamp shifts in each images.
Key functions include:
- \code{\link{extract_all_metadata}}: Extracts metadata and color statistics from JPEG images.
- \code{\link{calcColorSumStats}}: Computes grid-based color summaries.
- \code{\link{getGridPoints}}: Generates grid coordinates for image analysis.
}
\details{
Example usage:
\dontrun{
img_dir <- system.file("extdata", "BlackWhite", package = "CTTimeShift")
result <- extract_all_metadata(img_dir)
head(result)
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bitbucket.org/wegmannlab/timeshift/wiki/Home}
}

}
\author{
\strong{Maintainer}: Ernest Dadis Bush Fotsing \email{ernest.fotsing@unifr.ch}

Authors:
\itemize{
  \item Daniel Wegmann \email{daniel.wegmann@unifr.ch}
}

Other contributors:
\itemize{
  \item Aimée Freiberg [contributor]
}

}
\keyword{internal}
