% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT_timeshift.R
\name{extract_all_metadata}
\alias{extract_all_metadata}
\title{Extract Metadata and Color Statistics from Camera Trap Images}
\usage{
extract_all_metadata(root_dir)
}
\arguments{
\item{root_dir}{root_dir this the main directory where all the folder containg the image are located}
}
\value{
return a dataframe with  9 colum as follow: ImagesPath, Datetime, classification, is.color, dark_0.1, darrk_0.2, dark_0.3, dark_0.4, dark_0.5, proportion of black and white or color for each imagesy
}
\description{
Extract Metadata and Color Statistics from Camera Trap Images
}
\examples{
img_dir <- system.file("extdata", "BlackWhite", package = "CTTimeShift")
result <- extract_all_metadata(img_dir)
head(result)

}
\keyword{internal}
