% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT_timeshift.R
\name{getGridPoints}
\alias{getGridPoints}
\title{Function to create grid point on camera trap image}
\usage{
getGridPoints(picture, nrow = 9, ncol = 16)
}
\arguments{
\item{picture}{picture is the image of animal taken by camera trap}

\item{nrow}{A number of row of our grid that we want to set in the image that correspond to my grid}

\item{ncol}{A number of colum of our grid that we want to set in the image that correspond to my grid}
}
\value{
return a grid with coordinate x and y
}
\description{
Function to create grid point on camera trap image
}
\keyword{internal}
